+++ src/DTS/eBaySDK/ResolutionCaseManagement/Types/IssueFullRefundOptionType.php
@@ -23,7 +23,7 @@ namespace DTS\eBaySDK\ResolutionCaseManagement\Types;
 
 /**
  *
- * @property \DTS\eBaySDK\ResolutionCaseManagement\Types\Integer $daysToRefundBuyer
+ * @property integer $daysToRefundBuyer
  */
 class IssueFullRefundOptionType extends \DTS\eBaySDK\ResolutionCaseManagement\Types\BaseActivityOptionType
 {
@@ -32,7 +32,7 @@ class IssueFullRefundOptionType extends \DTS\eBaySDK\ResolutionCaseManagement\Ty
      */
     private static $propertyTypes = array(
         'daysToRefundBuyer' => array(
-            'type' => 'DTS\eBaySDK\ResolutionCaseManagement\Types\Integer',
+            'type' => 'integer',
             'unbound' => false,
             'attribute' => false,
             'elementName' => 'daysToRefundBuyer'